+++ contracts/tests/TestEAS.sol
@@ -7,21 +7,21 @@ import { EAS } from "../EAS.sol";
 import { ISchemaRegistry } from "../ISchemaRegistry.sol";
 
 contract TestEAS is EAS {
-    uint32 private constant INITIAL_TIME = 0;
+    uint64 private constant INITIAL_TIME = 0;
 
-    uint32 private _currentTime = INITIAL_TIME;
+    uint64 private _currentTime = INITIAL_TIME;
 
     constructor(ISchemaRegistry registry) EAS(registry) {}
 
-    function setTime(uint32 newTime) external {
+    function setTime(uint64 newTime) external {
         _currentTime = newTime;
     }
 
-    function getTime() external view returns (uint32) {
+    function getTime() external view returns (uint64) {
         return _time();
     }
 
-    function _time() internal view virtual override returns (uint32) {
+    function _time() internal view virtual override returns (uint64) {
         return _currentTime == INITIAL_TIME ? super._time() : _currentTime;
     }
 }
